+++ modules/builtin/src/actions/getNumberOfConversations.js
@@ -6,10 +6,8 @@
  * @param {string} output - The state variable to output the count to
  */
 const getNumberOfConversations = async output => {
-  const userId = event.target
-  const botId = event.botId
-  const key = bp.kvs.getUserStorageKey(userId, 'numberOfConversations')
-  const value = await bp.kvs.getStorageWithExpiry(botId, key)
+  const key = bp.kvs.getUserStorageKey(event.target, 'numberOfConversations')
+  const value = await bp.kvs.forBot(event.botId).getStorageWithExpiry(key)
 
   temp[output] = value
 }